Fix migration to avoid errors

+++ database/migrations/2025_06_04_195527_create_questions_table.php
@@ -11,22 +11,25 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('questions', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('question_bank_id')->constrained()->onDelete('cascade');
-            $table->foreignId('question_type_id')->constrained();
-            $table->text('question_text');
-            $table->json('metadata')->nullable(); // For additional configurations
-            $table->integer('points')->default(1);
-            $table->boolean('is_active')->default(true);
-            $table->foreignId('subject_id')->nullable();
-            $table->foreignId('academic_level_id')->nullable();
-            $table->foreignId('year_id')->nullable();
-            $table->integer('difficulty_level')->nullable();
-            $table->string('topic')->nullable();
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('questions')){
+            Schema::create('questions', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('question_bank_id')->constrained()->onDelete('cascade');
+                $table->foreignId('question_type_id')->constrained();
+                $table->text('question_text');
+                $table->json('metadata')->nullable(); // For additional configurations
+                $table->integer('points')->default(1);
+                $table->boolean('is_active')->default(true);
+                $table->foreignId('subject_id')->nullable();
+                $table->foreignId('academic_level_id')->nullable();
+                $table->foreignId('year_id')->nullable();
+                $table->integer('difficulty_level')->nullable();
+                $table->string('topic')->nullable();
+                $table->softDeletes();
+                $table->timestamps();
+            });
+        }
+        
     }
 
     /**
