+++ database/migrations/2025_06_04_232355_create_user_quiz_responses_table.php
@@ -11,16 +11,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('user_quiz_responses', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('attempt_id')->constrained('user_quiz_attempts')->onDelete('cascade');
-            $table->foreignId('question_id')->constrained()->onDelete('cascade');
-            $table->foreignId('option_id')->nullable()->constrained('options')->onDelete('cascade');
-            $table->text('answer_text')->nullable();
-            $table->boolean('is_correct')->default(false);
-            $table->integer('points_earned')->default(0);
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('user_quiz_responses')){
+            Schema::create('user_quiz_responses', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('attempt_id')->constrained('user_quiz_attempts')->onDelete('cascade');
+                $table->foreignId('question_id')->constrained()->onDelete('cascade');
+                $table->foreignId('option_id')->nullable()->constrained('options')->onDelete('cascade');
+                $table->text('answer_text')->nullable();
+                $table->boolean('is_correct')->default(false);
+                $table->integer('points_earned')->default(0);
+                $table->timestamps();
+            });
+        }
+
     }
 
     /**
